http2: fix potential warning about narrowing
By passing both arguments as qint64, then casting to qint32.
Since one of the arguments are qint32, the result will fit in qint32.

+++ src/network/access/qhttp2protocolhandler.cpp
@@ -504,7 +504,7 @@ bool QHttp2ProtocolHandler::sendDATA(Stream &stream)
         }
 
         frameWriter.start(FrameType::DATA, FrameFlag::EMPTY, stream.streamID);
-        const qint32 bytesWritten = std::min<qint32>(slot, chunkSize);
+        const qint32 bytesWritten = qint32(std::min<qint64>(slot, chunkSize));
 
         if (!frameWriter.writeDATA(*m_socket, maxFrameSize, src, bytesWritten))
             return false;
